Move `test_shrink_to_unwind` to its own file.
This way, no other test can be tripped up by `test_shrink_to_unwind` changing the alloc error hook.
